Responsive design changes
The responsive design for phone and tablet was modified the logo was centered and the menuan and othe icon were moved to imporve the view, the search bar for this design was created and functional. The logo was changed to an image png. Also changed the icons from svg to icons from boxicons. The login button works and saves everything in a json file.